ToggleButton fertig.

+++ HomeBacon/homebaconcore/src/main/java/de/whs/homebaconcore/WatchConnector.java
@@ -6,6 +6,6 @@ package de.whs.homebaconcore;
 public interface WatchConnector {
 
     void sendNote(Note note);
-    void startScan(int roomId);
+    void startScan(long roomId);
     void stopScan();
 }
